feat:Adding feature in checkout process for cart and direct buy.
Adding a new feature in checkout process for both cart checkout and direct
buy.It is determined by the Search params ProductId. Based on the ProductId
we send the products details to backend and also for the handling some logic
in the frontend.I also added the features for session expired when the checkout
is completed and user will notify they didn't go back to checkout process.User
did not go to checkout process directly they have to follow steps of checkout process.